Add --randomize flag, document missing flags, and enhance split screen/climax burst
- Add --randomize CLI flag: randomly enables effects, transitions, split screen,
  and climax burst for visual variety between runs; explicit flags take priority.
  Prints a fully-specified reproduce command before processing starts.
- Fix README: document --split-screen, --split-screen-count, --split-screen-panels,
  --climax-burst, --fast, and --randomize with examples and options table entries.
- Add double-time mode to climax burst (DOUBLE_TIME_PROBABILITY=1.0 default) and
  random split screens (50% chance); halves step duration for a faster pulse.
- Apply random per-panel effects in split screens and climax burst at PANEL_EFFECT_RATE.
- Fix uint8 cast in zoom_in effect to prevent PIL dtype error with ColorClip frames.
- Add tests for double-time durations, panel effects, and new constants.

+++ src/radiostar_killer/effects.py
@@ -78,7 +78,7 @@ def _make_zoom_in(rng: random.Random, clip: VideoClip) -> VideoClip:
         y_start = (h - new_h) // 2
 
         cropped = frame[y_start : y_start + new_h, x_start : x_start + new_w]
-        img = Image.fromarray(cropped)
+        img = Image.fromarray(cropped.astype(np.uint8))
         resized = img.resize((w, h), Image.Resampling.LANCZOS)
         return np.asarray(resized)
